modified optimal.cpp

+++ optimal.cpp
@@ -5,7 +5,7 @@
 
 using namespace std;
 
-void optimal(opt &ret, vector<vector<int>>& mydata, vector<float>& weights) {
+void optimal(opt &ret, vector<vector<float>>& mydata, vector<float>& weights) {
 
 	vector<float> data = mydata[0];
 	vector<int> index(data.size());
